Change method for setting candle color
Previously, the flickering algorithm was choosing a random vector in the (Red, Green) space. This caused the LED to get a green tint when the green LED was on but not the red.
The new color basis vectors are (Red, Yellow), which are then transformed into RGB, so that a naked green LED is never displayed.